// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/Cycai_adoptionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cycai_adoptionscycai_adoptionstatus = {
  100000000: 'Exploring',
  100000001: 'Implementing',
  100000002: 'Using'
} as const;
export type Cycai_adoptionscycai_adoptionstatus = keyof typeof Cycai_adoptionscycai_adoptionstatus;
export const Cycai_adoptionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cycai_adoptionsstatecode = keyof typeof Cycai_adoptionsstatecode;
export const Cycai_adoptionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cycai_adoptionsstatuscode = keyof typeof Cycai_adoptionsstatuscode;

export interface Cycai_adoptionsBase {
  cycai_adoptionid: string;
  cycai_adoptionstatus: Cycai_adoptionscycai_adoptionstatus;
  cycai_completedon?: string;
  cycai_name: string;
  cycai_projectname: string;
  cycai_solutionid: number;
  "cycai_startedbyid@odata.bind"?: string;
  cycai_startedon?: string;
  cycai_team?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cycai_adoptionsstatecode;
  statuscode?: Cycai_adoptionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cycai_adoptions extends Cycai_adoptionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cycai_adoptionstatusname?: string;
  cycai_startedbyidname?: string;
  cycai_startedbyidyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  cycai_startedbyid?: object;
  _cycai_startedbyid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}