// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/Cycai_participationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cycai_participationscycai_participationstatus = {
  100000000: 'Proposed',
  100000001: 'Accepted',
  100000002: 'Rejected',
  100000003: 'Withdrawn'
} as const;
export type Cycai_participationscycai_participationstatus = keyof typeof Cycai_participationscycai_participationstatus;
export const Cycai_participationscycai_targettype = {
  100000000: 'Idea',
  100000001: 'Solution'
} as const;
export type Cycai_participationscycai_targettype = keyof typeof Cycai_participationscycai_targettype;
export const Cycai_participationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cycai_participationsstatecode = keyof typeof Cycai_participationsstatecode;
export const Cycai_participationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cycai_participationsstatuscode = keyof typeof Cycai_participationsstatuscode;

export interface Cycai_participationsBase {
  "cycai_decidedbyid@odata.bind"?: string;
  cycai_decidedon?: string;
  cycai_message?: string;
  cycai_name: string;
  cycai_participationid: string;
  cycai_participationstatus: Cycai_participationscycai_participationstatus;
  cycai_rationale?: string;
  "cycai_requestedbyid@odata.bind"?: string;
  cycai_targetid: number;
  cycai_targetkey: string;
  cycai_targettype: Cycai_participationscycai_targettype;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cycai_participationsstatecode;
  statuscode?: Cycai_participationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cycai_participations extends Cycai_participationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cycai_decidedbyidname?: string;
  cycai_decidedbyidyominame?: string;
  cycai_participationstatusname?: string;
  cycai_requestedbyidname?: string;
  cycai_requestedbyidyominame?: string;
  cycai_targettypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  cycai_decidedbyid?: object;
  _cycai_decidedbyid_value?: string;
  cycai_requestedbyid?: object;
  _cycai_requestedbyid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}